Time: 7 ms (94.23%), Space: 45.2 MB (87.2%) - LeetHub

+++ 1885-count-number-of-homogenous-substrings/1885-count-number-of-homogenous-substrings.java
@@ -13,7 +13,7 @@ class Solution {
             count += ((w * (w+1)) / 2);
          }
         if (count > Integer.MAX_VALUE){
-            return (int)(count % (int)(Math.pow(10,9) + 7));
+            return (int)(count % (Math.pow(10,9) + 7));
         }else return (int)count;
     }
 }
